+++ ethereum/spec/src/testFixtures/java/tech/pegasys/teku/spec/propertytest/suppliers/blocks/versions/bellatrix/BlindedBeaconBlockBodySupplier.java
@@ -11,14 +11,16 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.spec.propertytest.suppliers.blocks;
+package tech.pegasys.teku.spec.propertytest.suppliers.blocks.versions.bellatrix;
 
+import tech.pegasys.teku.spec.SpecMilestone;
 import tech.pegasys.teku.spec.datastructures.blocks.blockbody.BeaconBlockBody;
 import tech.pegasys.teku.spec.propertytest.suppliers.DataStructureUtilSupplier;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
 
 public class BlindedBeaconBlockBodySupplier extends DataStructureUtilSupplier<BeaconBlockBody> {
+
   public BlindedBeaconBlockBodySupplier() {
-    super(DataStructureUtil::randomBlindedBeaconBlockBody);
+    super(DataStructureUtil::randomBlindedBeaconBlockBody, SpecMilestone.BELLATRIX);
   }
 }
